Continue to work on MTP support
Samsung MTP uses unknown functions that TWRP cannot define and understand so MTP fails to write to and from. It is currenty read only.

It seems the only way to get MTP to work fully is to merge Googles MTP source into the Samsung kernel source succesfully.

Using googles vice samsungs should allow mtp to write from device to pc or vice versa.

Known bugs as of this commit is MTP not allowing to write and vibration control does not work.